feat: implement FormHelpers module and integrate into Pathogen components
